LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-duration-auto.tentative-expected.txt


FAIL A value of auto can be specified for animation-duration assert_equals: expected "0" but got "-1"
FAIL e.style['animation-duration'] = "auto" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL Property animation-duration value 'auto' assert_true: 'auto' is a supported value for animation-duration. expected true got false
FAIL e.style['animation'] = "auto cubic-bezier(0, -2, 1, 3) -3s 4 reverse both paused anim" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL Property animation value 'auto cubic-bezier(0, -2, 1, 3) -3s 4 reverse both paused anim' assert_true: 'auto cubic-bezier(0, -2, 1, 3) -3s 4 reverse both paused anim' is a supported value for animation. expected true got false
